LayoutTests/media/media-session/user-gesture-action-handlers-expected.txt


RUN(internals.setMediaElementRestrictions(video, "requireusergestureforvideoratechange"))
RUN(video.src = findMediaFile("video", "../content/test"))
EVENT(canplay)
RUN(promise = video.play())
Promise rejected correctly OK
RUN(internals.sendMediaSessionAction(navigator.mediaSession, {action: "play"}))
ACTION: play
RUN(promise = video.play())
Promise resolved OK
END OF TEST
